backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.12
openai==1.52.2
python-dotenv==1.0.1